+++ changelog.md
@@ -1,10 +1,12 @@
 # EAS Python client
 ## [0.20.0] - UNRELEASED
 ### Breaking Changes
-* None.
+* `FixedTimeLoadOverride` now takes in optional list of floats instead of optional float for its variable.
 
 ### New Features
-* None.
+* Modification to `ModelConfig` to allow more customization for model generation
+  * Added support for separate `vMinPu` and `vMaxPu` with load and generators.
+  * Added support for `ctPrimScalingFactor` which is required when calculating new `ctPrim` value when feeder proxy loads are not in use.
 
 ### Enhancements
 * None.